+++ Arduino/HandTieArduinoSystem/SGManager.h
@@ -43,8 +43,8 @@ public:
    void allCalibration();
 
    void manualAssignPotPosForOneGauge(uint8_t, uint8_t, uint8_t);
-   void manualAssignTargetValNoAmpForOneGauge(uint8_t, uint8_t);
-   void manualAssignTargetValWithAmpForOneGauge(uint8_t, uint8_t);
+   void manualAssignTargetValNoAmpForOneGauge(uint8_t, uint16_t);
+   void manualAssignTargetValWithAmpForOneGauge(uint8_t, uint16_t);
 
 private:
    AnalogMux * analogMux;
